naivas/kfc

+++ naivas.py
@@ -116,7 +116,7 @@ Writing Data to Naivas Blob
                                   data=data)
 """
 
-def run_naivas_tasks_in_parallel(tasks):
+def run_naivas_tasks(tasks):
     running_tasks = [Process(target=task) for task in tasks]
     for running_task in running_tasks:
         running_task.start()
@@ -163,6 +163,8 @@ def nbo_task():
                 price = "-"
                 address = (soup.find('div', class_='header-user-address__content__text').text.strip())
                 status = 'unavailable'
+            if path.isfile(filename) is False:
+                raise Exception("File not found")
             with open(filename) as fp:
                 blob_products = json.load(fp)
             blob_products.append({ 'city': 'NBO', 'date': currentdatetime.strftime("%b %d, %Y"), 'time': currentdatetime.strftime("%H:00"), 'item': item, 'price': price, 'promo': promo, 'address': address, 'status' : status })
@@ -206,6 +208,8 @@ def nak_task():
                 price = "-"
                 address = (soup.find('div', class_='header-user-address__content__text').text.strip())
                 status = 'unavailable'
+            if path.isfile(filename) is False:
+                raise Exception("File not found")
             with open(filename) as fp:
                 blob_products = json.load(fp)
             blob_products.append({ 'city': 'NAK', 'date': currentdatetime.strftime("%b %d, %Y"), 'time': currentdatetime.strftime("%H:00"), 'item': item, 'price': price, 'promo': promo, 'address': address, 'status' : status })
@@ -249,6 +253,8 @@ def mbs_task():
                 price = "-"
                 address = (soup.find('div', class_='header-user-address__content__text').text.strip())
                 status = 'unavailable'
+            if path.isfile(filename) is False:
+                raise Exception("File not found")
             with open(filename) as fp:
                 blob_products = json.load(fp)
             blob_products.append({ 'city': 'MBS', 'date': currentdatetime.strftime("%b %d, %Y"), 'time': currentdatetime.strftime("%H:00"), 'item': item, 'price': price, 'promo': promo, 'address': address, 'status' : status })
@@ -292,6 +298,8 @@ def nrk_task():
                 price = "-"
                 address = (soup.find('div', class_='header-user-address__content__text').text.strip())
                 status = 'unavailable'
+            if path.isfile(filename) is False:
+                raise Exception("File not found")
             with open(filename) as fp:
                 blob_products = json.load(fp)
             blob_products.append({ 'city': 'NRK', 'date': currentdatetime.strftime("%b %d, %Y"), 'time': currentdatetime.strftime("%H:00"), 'item': item, 'price': price, 'promo': promo, 'address': address, 'status' : status })
@@ -335,6 +343,8 @@ def eld_task():
                 price = "-"
                 address = (soup.find('div', class_='header-user-address__content__text').text.strip())
                 status = 'unavailable'
+            if path.isfile(filename) is False:
+                raise Exception("File not found")
             with open(filename) as fp:
                 blob_products = json.load(fp)
             blob_products.append({ 'city': 'ELD', 'date': currentdatetime.strftime("%b %d, %Y"), 'time': currentdatetime.strftime("%H:00"), 'item': item, 'price': price, 'promo': promo, 'address': address, 'status' : status })
@@ -378,6 +388,8 @@ def ksm_task():
                 price = "-"
                 address = (soup.find('div', class_='header-user-address__content__text').text.strip())
                 status = 'unavailable'
+            if path.isfile(filename) is False:
+                raise Exception("File not found")
             with open(filename) as fp:
                 blob_products = json.load(fp)
             blob_products.append({ 'city': 'KSM', 'date': currentdatetime.strftime("%b %d, %Y"), 'time': currentdatetime.strftime("%H:00"), 'item': item, 'price': price, 'promo': promo, 'address': address, 'status' : status })
@@ -421,6 +433,8 @@ def thk_task():
                 price = "-"
                 address = (soup.find('div', class_='header-user-address__content__text').text.strip())
                 status = 'unavailable'
+            if path.isfile(filename) is False:
+                raise Exception("File not found")
             with open(filename) as fp:
                 blob_products = json.load(fp)
             blob_products.append({ 'city': 'THK', 'date': currentdatetime.strftime("%b %d, %Y"), 'time': currentdatetime.strftime("%H:00"), 'item': item, 'price': price, 'promo': promo, 'address': address, 'status' : status })
@@ -464,6 +478,8 @@ def syo_task():
                 price = "-"
                 address = (soup.find('div', class_='header-user-address__content__text').text.strip())
                 status = 'unavailable'
+            if path.isfile(filename) is False:
+                raise Exception("File not found")
             with open(filename) as fp:
                 blob_products = json.load(fp)
             blob_products.append({ 'city': 'SYO', 'date': currentdatetime.strftime("%b %d, %Y"), 'time': currentdatetime.strftime("%H:00"), 'item': item, 'price': price, 'promo': promo, 'address': address, 'status' : status })
@@ -507,20 +523,21 @@ def dia_task():
                 price = "-"
                 address = (soup.find('div', class_='header-user-address__content__text').text.strip())
                 status = 'unavailable'
+            if path.isfile(filename) is False:
+                raise Exception("File not found")
             with open(filename) as fp:
                 blob_products = json.load(fp)
             blob_products.append({ 'city': 'DIA', 'date': currentdatetime.strftime("%b %d, %Y"), 'time': currentdatetime.strftime("%H:00"), 'item': item, 'price': price, 'promo': promo, 'address': address, 'status' : status })
             with open(filename, 'w') as json_file: json.dump(blob_products, json_file)
 
-if __name__ == '__main__':
-    run_naivas_tasks_in_parallel([
-        nbo_task,
-        nak_task,
-        mbs_task,
-        nrk_task,
-        eld_task,
-        ksm_task,
-        thk_task,
-        syo_task,
-        dia_task,
-    ])
+run_naivas_tasks([
+    nbo_task,
+    nak_task,
+    mbs_task,
+    nrk_task,
+    eld_task,
+    ksm_task,
+    thk_task,
+    syo_task,
+    dia_task,
+])
